refactor: reorganize GUI for world
