--- src/views/world/view_water.cpp
@@ -1,10 +1,0 @@
-#include "fiber_pool.hpp"
-#include "views/view.hpp"
-
-namespace big
-{
-	void view::water()
-	{
-		components::command_checkbox<"partwater">();
-	}
-}
